requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
bcrypt==4.0.1
python-jwt
python-jose[cryptography]
langchain-pinecone
python-dotenv
langchain-core
langchain
langchain-ollama
langchain-community
langchain_groq
langchain_huggingface
langchain_openai
sentence_transformers
sse-starlette
einops
psycopg2-binary
tenacity